+++ utils/data/games.json
@@ -1,1 +1,1 @@
-[{"id":1,"title":"Crystal Keeper","description":"Управляй боевым дроном, чтобы любой ценой защитить кристалл от враждебных космослизней.","image":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/game/index.html"},{"id":2,"title":"Dangeons'n'Caves. Prologue","description":"Безымянный герой исследует пещеры и подземелья, чтобы найти больше информации о себе.","image":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/game/index.html"},{"id":3,"title":"123","image":"https://wiki-cdn.lesta.ru/images/0/05/GamePromo_World_of_Tanks.png","link":"http://sedfsdfsdf.asd","description":"asfd"}]
+[{"id":1,"title":"Crystal Keeper","description":"Управляй боевым дроном, чтобы любой ценой защитить кристалл от враждебных космослизней.","image":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/cristal-keeper/game/index.html"},{"id":2,"title":"Dangeons'n'Caves. Prologue","description":"Безымянный герой исследует пещеры и подземелья, чтобы найти больше информации о себе.","image":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/cover.png","link":"https://code.s3.yandex.net/teens/pindie-games/dangeons-n-caves-prologue/game/index.html"},{"id":3,"title":"тетрис","image":"https://avatars.mds.yandex.net/get-games/1892995/2a00000179a27a1f636db13680b60a43045c/orig","link":"http://sedfsdfsdf.asd","description":"что-то на умном"}]
